Adds object help response packets
