+++ Projects/Server.Tests/Tests/Network/Packets/Outgoing/MessagePacketTests.cs
@@ -184,6 +184,21 @@ namespace Server.Tests.Network
             AssertThat.Equal(result.Buffer[0].AsSpan(0), expected);
         }
 
+        [Fact]
+        public void TestObjectHelpResponse()
+        {
+            Serial s = 0x100;
+            var text = "This is some testing text";
+
+            var expected = new ObjectHelpResponse(s, text).Compile();
+
+            using var ns = PacketTestUtilities.CreateTestNetState();
+            ns.SendHelpResponse(s, text);
+
+            var result = ns.SendPipe.Reader.TryRead();
+            AssertThat.Equal(result.Buffer[0].AsSpan(0), expected);
+        }
+
         internal class TestPrompt : Prompt
         {
         }
